+++ src/Mzinga.Viewer/ViewModels/MainViewModel.cs
@@ -32,11 +32,11 @@ namespace Mzinga.Viewer.ViewModels
 
                 if (IsReviewMode)
                 {
-                    string fileName = AppVM.EngineWrapper.CurrentGameSettings?.GameRecording?.FileName;
+                    var fileUri = AppVM.EngineWrapper.CurrentGameSettings?.GameRecording?.FileUri;
 
-                    if (!string.IsNullOrWhiteSpace(fileName))
+                    if (fileUri is not null)
                     {
-                        title = $"{fileName} - {title}";
+                        title = $"{(fileUri.IsFile ? fileUri.LocalPath : fileUri.ToString())} - {title}";
                     }
                 }
 
@@ -451,11 +451,11 @@ namespace Mzinga.Viewer.ViewModels
                                 AppVM.EngineWrapper.CurrentGameSettings.Metadata.Clear();
                                 AppVM.EngineWrapper.CurrentGameSettings.Metadata.CopyFrom(metadata);
 
-                                StrongReferenceMessenger.Default.Send(new SaveGameMessage(AppVM.EngineWrapper.CurrentGameSettings.GameRecording, (fileName) =>
+                                StrongReferenceMessenger.Default.Send(new SaveGameMessage(AppVM.EngineWrapper.CurrentGameSettings.GameRecording, (fileUri) =>
                                 {
-                                    if (IsReviewMode)
+                                    if (fileUri is not null && IsReviewMode)
                                     {
-                                        AppVM.EngineWrapper.CurrentGameSettings.GameRecording.FileName = fileName;
+                                        AppVM.EngineWrapper.CurrentGameSettings.GameRecording.FileUri = fileUri;
                                         OnPropertyChanged(nameof(Title));
                                     }
                                 }));