+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Viewer: Switch between Light and Dark themes without restarting the app
 * Viewer: Fixed "crash" when exiting while Internal Engine is still running
+* Viewer: Fixed issues when loading newer BoardSpace SGF files
 * Viewer: Minor rendering improvements
 * Viewer: Switched to CompiledBindings for improved performance
 * Viewer: Updated New Game Window